+++ src/main/java/nextstep/auth/AuthenticationPrincipalArgumentResolver.java
@@ -1,5 +1,6 @@
 package nextstep.auth;
 
+import nextstep.auth.util.AuthorizationExtractor;
 import nextstep.support.AuthorizationException;
 import org.springframework.core.MethodParameter;
 import org.springframework.web.bind.support.WebDataBinderFactory;
@@ -10,6 +11,11 @@ import org.springframework.web.method.support.ModelAndViewContainer;
 import javax.servlet.http.HttpServletRequest;
 
 public class AuthenticationPrincipalArgumentResolver implements HandlerMethodArgumentResolver {
+    private JwtTokenProvider jwtTokenProvider;
+
+    public AuthenticationPrincipalArgumentResolver(JwtTokenProvider jwtTokenProvider) {
+        this.jwtTokenProvider = jwtTokenProvider;
+    }
 
     @Override
     public boolean supportsParameter(MethodParameter parameter) {
@@ -19,11 +25,12 @@ public class AuthenticationPrincipalArgumentResolver implements HandlerMethodArg
     @Override
     public Object resolveArgument(MethodParameter parameter, ModelAndViewContainer mavContainer, NativeWebRequest webRequest, WebDataBinderFactory binderFactory) {
         HttpServletRequest request = webRequest.getNativeRequest(HttpServletRequest.class);
-        String headerValue = AuthorizationExtractor.extract(request);
-        if (headerValue == null) {
+        String token = AuthorizationExtractor.extract(request);
+        if (token == null) {
             throw new AuthorizationException();
         }
-        String payload = new JwtTokenProvider().getPrincipal(headerValue);
+        jwtTokenProvider.validateToken(token);
+        String payload = jwtTokenProvider.getPrincipal(token);
         if (payload == null) {
             throw new AuthorizationException();
         }
