+++ src/main/java/nextstep/auth/util/AuthorizationExtractor.java
@@ -1,4 +1,4 @@
-package nextstep.auth;
+package nextstep.auth.util;
 
 import javax.servlet.http.HttpServletRequest;
 import java.util.Enumeration;
