+++ src/main/java/nextstep/config/WebMvcConfiguration.java
@@ -1,5 +1,7 @@
-package nextstep.auth;
+package nextstep.config;
 
+import nextstep.auth.AuthenticationPrincipalArgumentResolver;
+import nextstep.auth.JwtTokenProvider;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
@@ -8,9 +10,15 @@ import java.util.List;
 
 @Configuration
 public class WebMvcConfiguration implements WebMvcConfigurer {
+    private JwtTokenProvider jwtTokenProvider;
+
+    public WebMvcConfiguration(JwtTokenProvider jwtTokenProvider) {
+        this.jwtTokenProvider = jwtTokenProvider;
+    }
+
     @Override
     public void addArgumentResolvers(List<HandlerMethodArgumentResolver> resolvers) {
-        resolvers.add(new AuthenticationPrincipalArgumentResolver());
+        resolvers.add(new AuthenticationPrincipalArgumentResolver(jwtTokenProvider));
     }
 }
